Move all this editor logic out of Editor component
